App.js: untangle code.
