+++ src/Components/StaticHeader/useStaticHeader.js
@@ -1,0 +1,15 @@
+import { useState } from "react";
+
+const useStaticHeader = () => {
+  const [staticHeader, setStaticHeader] = useState({
+    title: "Loading...",
+    onBackButtonClick: null,
+  });
+
+  return {
+    staticHeader,
+    setStaticHeader,
+  };
+};
+
+export default useStaticHeader;
